CPlot: portage i8, OCC: portage i8
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4276 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ CPlot/CPlot/Display/displaySIsoSolid.cpp
@@ -111,7 +111,7 @@ void DataDL::displaySIsoSolid()
       _shaders[s]->setUniform("colormap", (int)1);
       if (s == _shaders.shader_id(shader::vector_arrow))
       {
-        _shaders[s]->setUniform("show_surface", ptrState->vectorShowSurface);
+        _shaders[s]->setUniform("show_surface", (int)ptrState->vectorShowSurface);
         _shaders[s]->setUniform("project_vectors", (int)ptrState->vector_projection);
         _shaders[s]->setUniform("style_arrow", (int)ptrState->vectorShape);
       }
